ignoring /bin/ (eclipse) added ts3 manager to handle the ts3api, query and config clientDbId is Integer now Added basics of the ts3Login Module

+++ src/main/java/eu/kandru/luna/model/json/AuthChallengeRequest.java
@@ -18,5 +18,5 @@ public class AuthChallengeRequest {
     /**
      * The ID that identifies the client in the database.
      */
-    private Long clientDbId;
+    private Integer clientDbId;
 }
